requirements.txt
# Additional requirements go here
django-kernel